Check plugin version on activation of plugin (#6)
When we release a new version on the QGIS repository we want to notify the user that a new version is available. This enhancement is used to check that the GIS4WRF version installed on the system is the latest, else we prompt the user with a message asking to update to the latest version.